Moved fixtures again to submodule and added conftest.py.
