+++ tests/conftest.py
@@ -1,0 +1,7 @@
+"""Conftest with all pytest plugins for HyperSHAP."""
+
+from __future__ import annotations
+
+pytest_plugins = [
+    "tests.fixtures.simple_setup",
+]
